убрала лишнее в своем коде
