+++ src/js/popular.js
@@ -1,6 +1,6 @@
 //FT-07 Реализовать подгрузку популярных фильмов на главную (первую) страницу
 import { paginationPopularFilms } from './pagination';
-import getFilms from './fetch-popular';
+import getFilms from './fetch';
 // import AxiosApi from './fetch-popular'
 import gallery from '../templates/film-card.hbs';
 import genres from './genres.json';
